:recycle: (pc-15004) move logoutFromEduconnectIfAllowed
change in CSS snapshot does not lead to any change in-app
